if downloader failed, then don't try to copy the file
This avoids an exception from the shutil.  Returning here just means
the annotator doesn't start up, which is reasonably the best thing we
can do.

+++ clients/marker.py
@@ -864,6 +864,9 @@ class MarkerClient(QWidget):
                         return
                     count = 0
 
+        # maybe the downloader failed for some (rare) reason
+        if not os.path.exists(fname):
+            return
         print("Debug: original image {} copy to paperdir {}".format(fname, paperdir))
         shutil.copyfile(fname, aname)
